Code review by @raman-m
