+++ src/Ocelot/Configuration/File/FileCacheOptions.cs
@@ -12,25 +12,15 @@ public class FileCacheOptions
         EnableContentHashing = from.EnableContentHashing;
     }
 
-    /// <summary>
-    /// using int? to have null as default value
-    /// and allowing global configuration usage
-    /// If null then use global configuration with 0 by default.
-    /// </summary>
-    /// <value>
-    /// The time to live seconds, with 0 by default.
-    /// </value>
+    /// <summary>Using <see cref="Nullable{T}"/> where T is <see cref="int"/> to have <see langword="null"/> as default value and allowing global configuration usage.</summary>
+    /// <remarks>If <see langword="null"/> then use global configuration with 0 by default.</remarks>
+    /// <value>The time to live seconds, with 0 by default.</value>
     public int? TtlSeconds { get; set; }
     public string Region { get; set; }
     public string Header { get; set; }
 
-    /// <summary>
-    /// using bool? to have null as default value
-    /// and allowing global configuration usage
-    /// If null then use global configuration with false by default.
-    /// </summary>
-    /// <value>
-    /// True if content hashing is enabled; otherwise, false.
-    /// </value>
+    /// <summary>Using <see cref="Nullable{T}"/> where T is <see cref="bool"/> to have <see langword="null"/> as default value and allowing global configuration usage.</summary>
+    /// <remarks>If <see langword="null"/> then use global configuration with <see langword="false"/> by default.</remarks>
+    /// <value><see langword="true"/> if content hashing is enabled; otherwise, <see langword="false"/>.</value>
     public bool? EnableContentHashing { get; set; }
 }
